Create file-backup.bat script
This was a script that I created when my boss asked me to ensure a local PST file I was maintaining of a group email box was backed up to the network. I then modified the script to make it easier for others to plug in the values of what they may need to back up rather than digging into the depths of the script and making changes there.

I used portions of my previously created ordinal-date.bat script in this script.