Add missing escape characters, fix perllocal find command, and fix REQURIES flag
This fixes several variables that weren't properly escaped. This also changes the perllocal find command to match the cmake and perl template rather than the autotools version. Finally, this fixes the mixed variables for the REQUIRES flag so sbgen.sh -r "dep1 dep2" works again.

Thanks to Daniel Prosser for the reports.