transcode-265.sh: Add catch for bigger transcoded size and reformat output
Out of all the runs I've done, only one ended up bigger than when
I started. I figured I should add logic to catch and display if
this happens in the future.

I also took this time to reformat the output a bit and to save
the stats of the output at the output directory. This will allow
you to see the stats even if the shell output disappears.