transcode-265.sh: Fix final output to show increases and decreases
I forgot why I had two separate outputs when completing a conversion. I had one
for the most common result, a reduction in final size, but I had added another
one because I had a time where the resulting size was larger than the source.
When I moved the output to a function, I forgot that I needed to cover both
possibilities. This fixes that.