transcode-265.sh: Fix global stats output
When I added the functions for checking directories and files, I moved the
global stats initialization part of the script (which includes where it sources
the stats file) lower in the script, which ended up being lower than the first
help function calls. This fix is to just move those functions higher up in the
script to ensure that the global stats file has been sourced before we try and
run the help function.