sbgen.sh: Add interactive mode to prompt for information
Interactive mode will allow prompting for all inputs instead of requiring
passing commandline options to the script. During development,

I also noticed that perl scripts were not being properly formatted since there
is a separate SRCNAM function that wasn't initially included in my output.

Finally, I moved the script types to a function and made a variable of the
count of scripts to be used throughout the script (then I just need to change
a single variable when I add a new script instead of hunting throughout the
script).